+++ src/kernels/MIOpenScatterND.cpp
@@ -55,7 +55,7 @@ __device__ void ScatterNDAddForward_Kernel(const TIO* input,
 
     for(int dim = slice_dim - 1; dim >= 0; dim--)
     {
-        if(dim == slice_dim - 1)
+        if(static_cast<size_t>(dim) == slice_dim - 1)
         {
             batch_strides[dim] = 1;
         }
