+++ test/cpu_gather.hpp
@@ -52,7 +52,7 @@ void cpu_gathernd_backward(const tensor<T>& outputGrad,
 
     for(int dim = slice_dim - 1; dim >= 0; dim--)
     {
-        if(dim == slice_dim - 1)
+        if(static_cast<size_t>(dim) == slice_dim - 1)
         {
             batch_strides[dim] = 1;
         }
